Reformatted server into a separate module
