+++ cmd/main.go
@@ -1,10 +1,10 @@
 package main
 
 import (
+	"github.com/JJ-Intelligence/SR-Games-Backend/internal/server"
 	"log"
 	"net/http"
 	"os"
-	"github.com/JJ-Intelligence/SR-Games-Backend/pkg/server"
 )
 
 func main() {
